fix(sas): Use a ReusableSecret for the Sas curve25519 secret key
Methods that take self can't easily be bound over FFI, since the
intetion of this crate is to work in various languages let's switch to a
method that takes a borrow of self.

This is a bit unfortunate since we do not want people to reuse the
secret key here.